Refactored FGS and IMaGES (and their editors and all related files) so as to combine them into one algorithm and one editor. In the Tetrad interface GES has been removed (because it's buggy), and IMaGES has been removed, and both of them have been replaced by "FGS-IMaGES".
