+++ tetrad-gui/src/main/java/edu/cmu/tetradapp/editor/FactorAnalysisEditor.java
@@ -158,8 +158,8 @@ public class FactorAnalysisEditor extends AbstractSearchEditor {
                 return new GesIndTestParamsEditor(params, discreteData);
             }
 
-            if (getAlgorithmRunner() instanceof  ImagesRunner) {
-                ImagesRunner gesRunner = ((ImagesRunner) getAlgorithmRunner());
+            if (getAlgorithmRunner() instanceof FgsRunner) {
+                FgsRunner gesRunner = ((FgsRunner) getAlgorithmRunner());
                 GesIndTestParams params = (GesIndTestParams) indTestParams;
                 DataSet dataSet = (DataSet) gesRunner.getDataModel();
                 boolean discreteData = dataSet.isDiscrete();
