+++ tetrad-gui/src/main/java/edu/cmu/tetradapp/editor/PcGesSearchEditor.java
@@ -135,7 +135,7 @@ public class PcGesSearchEditor extends AbstractSearchEditor
         super(runner, "Result Pattern");
     }
 
-    public PcGesSearchEditor(ImagesRunner runner) {
+    public PcGesSearchEditor(FgsRunner runner) {
         super(runner, "Result Pattern");
     }
 
@@ -791,8 +791,8 @@ public class PcGesSearchEditor extends AbstractSearchEditor
                 return new GesIndTestParamsEditor(params, discreteData);
             }
 
-            if (getAlgorithmRunner() instanceof ImagesRunner) {
-                ImagesRunner gesRunner = ((ImagesRunner) getAlgorithmRunner());
+            if (getAlgorithmRunner() instanceof FgsRunner) {
+                FgsRunner gesRunner = ((FgsRunner) getAlgorithmRunner());
                 GesIndTestParams params = (GesIndTestParams) indTestParams;
                 DataSet dataSet = (DataSet) gesRunner.getDataModel();
                 boolean discreteData = dataSet.isDiscrete();
