+++ tetrad-gui/src/main/java/edu/cmu/tetradapp/model/FciIndTestParams.java
@@ -70,6 +70,7 @@ public class FciIndTestParams implements IndTestParams {
      */
     private int maxReachablePathLength = -1;
     private boolean gFci;
+    private int numPatternsToSave = 0;
 
     //============================CONSTRUCTORS=========================//
 
@@ -190,6 +191,15 @@ public class FciIndTestParams implements IndTestParams {
 
     public void setGFci(boolean gFci) {
         this.gFci = gFci;
+    }
+
+    @Override
+    public int getNumPatternsToSave() {
+        return numPatternsToSave;
+    }
+
+    public void setNumPatternsToSave(int numPatternsToSave) {
+        this.numPatternsToSave = numPatternsToSave;
     }
 }
