+++ tetrad-gui/src/main/java/edu/cmu/tetradapp/model/MergeDatasetsWrapper.java
@@ -87,6 +87,8 @@ public class MergeDatasetsWrapper extends DataWrapper {
     }
 
     private void construct(DataWrapper...dataWrappers) {
+        System.out.println("A");
+
         for (DataWrapper wrapper : dataWrappers) {
             if (wrapper == null) {
                 throw new NullPointerException("The given data must not be null");
@@ -96,15 +98,14 @@ public class MergeDatasetsWrapper extends DataWrapper {
         DataModelList merged = new DataModelList();
 
         for (DataWrapper wrapper : dataWrappers) {
-            for (DataModel model : wrapper.getDataModelList()) {
-                merged.add(model);
-            }
+            merged.addAll(wrapper.getDataModelList());
         }
 
         this.setDataModel(merged);
 
         LogDataUtils.logDataModelList("Parent data in which constant columns have been removed.", getDataModelList());
 
+        System.out.println("B");
 
     }
