+++ tetrad-gui/src/main/resources/resources/configpost.xml
@@ -1437,24 +1437,41 @@
                 <params-class>edu.cmu.tetradapp.model.BasicSearchParams
                 </params-class>
             </model>
-            <model name="GES" acronym="GES" help="ges"
-                   category="Pattern Searches">
-                <logger>
-                    <event id="info" description="Information" default="on"/>
-                    <event id="insertedEdges" description="Inserted Edges" default="on"/>
-                    <event id="deletedEdges" description="Deleted Edges" default="on"/>
-                    <event id="directedEdges" description="Directed Edges" default="on"/>
-                    <event id="edgeEvaluations" description="Edge Evaluations"/>
-                    <event id="impliedOrientations" description="Implied Orientations"/>
-                    <event id="rebuiltPatterns" description="Rebuilt Patterns"/>
-                    <event id="graph" description="Graph" default="on"/>
-                </logger>
-                <model-class>edu.cmu.tetradapp.model.GesRunner</model-class>
-                <editor-class>edu.cmu.tetradapp.editor.GesSearchEditor
-                </editor-class>
-                <params-class>edu.cmu.tetradapp.model.GesParams</params-class>
-            </model>
-            <model name="FGS" acronym="FGS" help="fgs"
+            <!--<model name="GES" acronym="GES" help="ges" category="Unlisted">-->
+                   <!--&lt;!&ndash;category="Pattern Searches">&ndash;&gt;-->
+                <!--<logger>-->
+                    <!--<event id="info" description="Information" default="on"/>-->
+                    <!--<event id="insertedEdges" description="Inserted Edges" default="on"/>-->
+                    <!--<event id="deletedEdges" description="Deleted Edges" default="on"/>-->
+                    <!--<event id="directedEdges" description="Directed Edges" default="on"/>-->
+                    <!--<event id="edgeEvaluations" description="Edge Evaluations"/>-->
+                    <!--<event id="impliedOrientations" description="Implied Orientations"/>-->
+                    <!--<event id="rebuiltPatterns" description="Rebuilt Patterns"/>-->
+                    <!--<event id="graph" description="Graph" default="on"/>-->
+                <!--</logger>-->
+                <!--<model-class>edu.cmu.tetradapp.model.GesRunner</model-class>-->
+                <!--<editor-class>edu.cmu.tetradapp.editor.GesSearchEditor-->
+                <!--</editor-class>-->
+                <!--<params-class>edu.cmu.tetradapp.model.GesParams</params-class>-->
+            <!--</model>-->
+            <!--<model name="FGS" acronym="FGS" help="fgs" category="Unlisted">-->
+                   <!--&lt;!&ndash;category="Pattern Searches">&ndash;&gt;-->
+                <!--<logger>-->
+                    <!--<event id="info" description="Information" default="on"/>-->
+                    <!--<event id="insertedEdges" description="Inserted Edges" default="on"/>-->
+                    <!--<event id="deletedEdges" description="Deleted Edges" default="on"/>-->
+                    <!--<event id="directedEdges" description="Directed Edges" default="on"/>-->
+                    <!--<event id="edgeEvaluations" description="Edge Evaluations"/>-->
+                    <!--<event id="impliedOrientations" description="Implied Orientations"/>-->
+                    <!--<event id="rebuiltPatterns" description="Rebuilt Patterns"/>-->
+                    <!--<event id="graph" description="Graph" default="on"/>-->
+                <!--</logger>-->
+                <!--<model-class>edu.cmu.tetradapp.model.FgsRunnerOld</model-class>-->
+                <!--<editor-class>edu.cmu.tetradapp.editor.FgsSearchEditor-->
+                <!--</editor-class>-->
+                <!--<params-class>edu.cmu.tetradapp.model.FgsParams</params-class>-->
+            <!--</model>-->
+            <model name="FGS-IMaGES" acronym="FGS" help="fgs"
                    category="Pattern Searches">
                 <logger>
                     <event id="info" description="Information" default="on"/>
@@ -1470,23 +1487,6 @@
                 <editor-class>edu.cmu.tetradapp.editor.FgsSearchEditor
                 </editor-class>
                 <params-class>edu.cmu.tetradapp.model.FgsParams</params-class>
-            </model>
-            <model name="iMAGES" acronym="iMAGES" help="ges"
-                   category="Pattern Searches">
-                <logger>
-                    <event id="info" description="Information" default="on"/>
-                    <event id="insertedEdges" description="Inserted Edges" default="on"/>
-                    <event id="deletedEdges" description="Deleted Edges" default="on"/>
-                    <event id="directedEdges" description="Directed Edges" default="on"/>
-                    <event id="edgeEvaluations" description="Edge Evaluations"/>
-                    <event id="impliedOrientations" description="Implied Orientations"/>
-                    <event id="rebuiltPatterns" description="Rebuilt Patterns"/>
-                    <event id="graph" description="Graph" default="on"/>
-                </logger>
-                <model-class>edu.cmu.tetradapp.model.ImagesRunner</model-class>
-                <editor-class>edu.cmu.tetradapp.editor.ImagesSearchEditor
-                </editor-class>
-                <params-class>edu.cmu.tetradapp.model.GesParams</params-class>
             </model>
             <!--Heuristic Best Significant Model Search-->
             <model name="BFF" acronym="BFF" help="bff"
