+++ tetrad-lib/src/main/java/edu/cmu/tetrad/data/CovarianceMatrix.java
@@ -415,7 +415,13 @@ public class CovarianceMatrix implements ICovarianceMatrix {
 
     public void setVariables(List<Node> variables) {
         if (variables.size() != this.variables.size()) throw new IllegalArgumentException("Wrong # of variables.");
-        this.variables = variables;
+        for (int i = 0; i < variables.size(); i++) {
+            if (!variables.get(i).getName().equals(variables.get(i).getName())) {
+                throw new IllegalArgumentException("Variable in index " + (i + 1) + " does not have the same name " +
+                        "as the variable being substituted for it.");
+            }
+            this.variables = variables;
+        }
     }
 
     @Override
