+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/Fgs.java
@@ -28,6 +28,8 @@ import edu.cmu.tetrad.util.ForkJoinPoolInstance;
 import edu.cmu.tetrad.util.TetradLogger;
 
 import java.io.PrintStream;
+import java.text.DecimalFormat;
+import java.text.NumberFormat;
 import java.util.*;
 import java.util.concurrent.*;
 
@@ -297,8 +299,8 @@ public final class Fgs implements GraphSearch, GraphScorer {
      * For BIC score, a multiplier on the penalty term. For continuous searches.
      */
     public double getPenaltyDiscount() {
-        if (gesScore instanceof SemBicScore) {
-            return ((SemBicScore) gesScore).getPenaltyDiscount();
+        if (gesScore instanceof ISemBicScore) {
+            return ((ISemBicScore) gesScore).getPenaltyDiscount();
         } else {
             return 2.0;
         }
@@ -308,13 +310,8 @@ public final class Fgs implements GraphSearch, GraphScorer {
      * For BIC score, a multiplier on the penalty term. For continuous searches.
      */
     public void setPenaltyDiscount(double penaltyDiscount) {
-        if (penaltyDiscount < 0) {
-            throw new IllegalArgumentException("Penalty penaltyDiscount must be >= 0: "
-                    + penaltyDiscount);
-        }
-
-        if (gesScore instanceof SemBicScore) {
-            ((SemBicScore) gesScore).setPenaltyDiscount(penaltyDiscount);
+        if (gesScore instanceof ISemBicScore) {
+            ((ISemBicScore) gesScore).setPenaltyDiscount(penaltyDiscount);
         }
     }
 
@@ -1561,12 +1558,7 @@ public final class Fgs implements GraphSearch, GraphScorer {
             }
 
             int yIndex = hashIndices.get(y);
-
-            for (int i = 0; i < parentIndices.length - 1; i++) {
-                int[] _parents = Arrays.copyOf(parentIndices, i);
-                int xIndex = parentIndices[i + 1];
-                score += gesScore.localScoreDiff(yIndex, _parents, xIndex);
-            }
+            score += gesScore.localScore(yIndex, parentIndices);
         }
 
         return score;
@@ -1603,6 +1595,60 @@ public final class Fgs implements GraphSearch, GraphScorer {
             topGraphs.removeFirst();
         }
     }
+
+    public String logEdgeBayesFactorsString(Graph dag) {
+        Map<Edge, Double> factors = logEdgeBayesFactors(dag);
+        return logBayesPosteriorFactorsString(factors, scoreDag(dag));
+    }
+
+    public Map<Edge, Double> logEdgeBayesFactors(Graph dag) {
+        Map<Edge, Double> logBayesFactors = new HashMap<Edge, Double>();
+        double withEdge = scoreDag(dag);
+
+        for (Edge edge : dag.getEdges()) {
+            dag.removeEdge(edge);
+            double withoutEdge = scoreDag(dag);
+            double difference = withEdge - withoutEdge;
+            logBayesFactors.put(edge, difference);
+            dag.addEdge(edge);
+        }
+
+        return logBayesFactors;
+    }
+
+    private String logBayesPosteriorFactorsString(final Map<Edge, Double> factors, double modelScore) {
+        NumberFormat nf = new DecimalFormat("0.00");
+        StringBuilder builder = new StringBuilder();
+
+        List<Edge> edges = new ArrayList<>(factors.keySet());
+
+        Collections.sort(edges, new Comparator<Edge>() {
+            @Override
+            public int compare(Edge o1, Edge o2) {
+                return -Double.compare(factors.get(o1), factors.get(o2));
+            }
+        });
+
+        builder.append("Edge Posterior Log Bayes Factors:\n\n");
+
+        builder.append("For a DAG in the IMaGES pattern with model score m, for each edge e in the " +
+                "DAG, the model score that would result from removing each edge, calculating " +
+                "the resulting model score m(e), and then reporting m(e) - m. The score used is " +
+                "the IMScore, L - SUM_i{kc ln n(i)}, L is the maximum likelihood of the model, " +
+                "k isthe number of parameters of the model, n(i) is the sample size of the ith " +
+                "data set, and c is the penalty penaltyDiscount. Note that the more negative the score, " +
+                "the more important the edge is to the posterior probability of the IMaGES model. " +
+                "Edges are given in order of their importance so measured.\n\n");
+
+        int i = 0;
+
+        for (Edge edge : edges) {
+            builder.append(++i).append(". ").append(edge).append(" ").append(nf.format(factors.get(edge))).append("\n");
+        }
+
+        return builder.toString();
+    }
+
 }
