+++ tetrad-lib/src/main/java/edu/cmu/tetrad/search/ISemBicScore.java
@@ -1,0 +1,9 @@
+package edu.cmu.tetrad.search;
+
+/**
+ * Created by jdramsey on 2/21/16.
+ */
+public interface ISemBicScore extends GesScore {
+    void setPenaltyDiscount(double penaltyDiscount);
+    double getPenaltyDiscount();
+}
